requirements.txt
anthropic>=0.96.0
python-dotenv>=1.2.2
python-telegram-bot[job-queue]>=22.7
faster-whisper>=1.0.0
PyAudio>=0.2.14
numpy>=1.24.0
